Add files via upload
This program is a University Personnel Information System that manages a linked list of records for students and professors. It allows you to add, update, list, and delete records, then saves the data to a specified file upon exit. It dynamically allocates memory for each record, uses file I/O to store the data, and processes commands to manipulate the records.